+++ ui/src/components/list-view/QasListView.vue
@@ -15,7 +15,7 @@
         </div>
 
         <div v-else-if="!mx_isFetching">
-          <slot v-if="mx_canShowErrorSlot" name="error" />
+          <slot v-if="mx_canShowFetchErrorSlot" name="fetch-error" />
 
           <slot v-else name="empty-results">
             <qas-empty-result-text />