+++ ui/src/composables/private/use-view.js
@@ -80,7 +80,7 @@ export default function useView (config) {
 
   // refs
   const cancelBeforeFetch = ref(false)
-  const isFetchError = ref(false)
+  const hasFetchError = ref(false)
 
   // constants
   const fetchErrorMessage = 'Ops… Não conseguimos acessar as informações. Por favor, tente novamente em alguns minutos.'
@@ -89,8 +89,8 @@ export default function useView (config) {
   const componentClass = computed(() => props.useBoundary && 'container spaced')
   const hasFooterSlot = computed(() => !!slots.footer)
   const hasHeaderSlot = computed(() => !!slots.header)
-  const hasErrorSlot = computed(() => !!slots.error)
-  const canShowErrorSlot = computed(() => hasErrorSlot.value && isFetchError.value)
+  const hasFetchErrorSlot = computed(() => !!slots.fetchError)
+  const canShowFetchErrorSlot = computed(() => hasFetchErrorSlot.value && hasFetchError.value)
 
   // watch
   watch(() => viewState.value.fetching, value => emit('update:fetching', value))
@@ -114,7 +114,7 @@ export default function useView (config) {
       return
     }
 
-    isFetchError.value = true
+    hasFetchError.value = true
 
     NotifyError(fetchErrorMessage)
   }
@@ -172,7 +172,7 @@ export default function useView (config) {
     componentClass,
     hasFooterSlot,
     hasHeaderSlot,
-    canShowErrorSlot,
+    canShowFetchErrorSlot,
 
     // functions
     errorHandler,